refactor: move GlobalError content to a separate component
